+++ stdb/query/BUILD
@@ -39,22 +39,22 @@ cc_library(
     "query_processing/spacesaver.h",
     "query_processing/top.h",
 
-    "plan/aggregate_combiner.h",
-    "plan/aggregate.h",
-    "plan/aggregate_processing_step.h",
-    "plan/chain.h",
-    "plan/filter_processing_step.h",
-    "plan/group_aggregate_combiner.h",
-    "plan/group_aggregate_filter_processing_step.h",
-    "plan/group_aggregate_processing_step.h",
-    "plan/join.h",
-    "plan/materialization_step.h",
-    "plan/mergeby.h",
-    "plan/processing_prelude.h",
-    "plan/scan_events_processing_step.h",
-    "plan/scan_processing_step.h",
-    "plan/series_order_aggregate.h",
-    "plan/time_order_aggregate.h",
+    "steps/aggregate_combiner.h",
+    "steps/aggregate.h",
+    "steps/aggregate_processing_step.h",
+    "steps/chain.h",
+    "steps/filter_processing_step.h",
+    "steps/group_aggregate_combiner.h",
+    "steps/group_aggregate_filter_processing_step.h",
+    "steps/group_aggregate_processing_step.h",
+    "steps/join.h",
+    "steps/materialization_step.h",
+    "steps/mergeby.h",
+    "steps/processing_prelude.h",
+    "steps/scan_events_processing_step.h",
+    "steps/scan_processing_step.h",
+    "steps/series_order_aggregate.h",
+    "steps/time_order_aggregate.h",
   ],
   alwayslink = 1,
   copts = [
