+++ stdb/query/queryplan.cc
@@ -3,22 +3,22 @@
  */
 #include "stdb/query/queryplan.h"
 
-#include "stdb/query/plan/aggregate_combiner.h"
-#include "stdb/query/plan/aggregate.h"
-#include "stdb/query/plan/aggregate_processing_step.h"
-#include "stdb/query/plan/chain.h"
-#include "stdb/query/plan/filter_processing_step.h"
-#include "stdb/query/plan/group_aggregate_combiner.h"
-#include "stdb/query/plan/group_aggregate_filter_processing_step.h"
-#include "stdb/query/plan/group_aggregate_processing_step.h"
-#include "stdb/query/plan/join.h"
-#include "stdb/query/plan/materialization_step.h"
-#include "stdb/query/plan/mergeby.h"
-#include "stdb/query/plan/processing_prelude.h"
-#include "stdb/query/plan/scan_events_processing_step.h"
-#include "stdb/query/plan/scan_processing_step.h"
-#include "stdb/query/plan/series_order_aggregate.h"
-#include "stdb/query/plan/time_order_aggregate.h"
+#include "stdb/query/steps/aggregate_combiner.h"
+#include "stdb/query/steps/aggregate.h"
+#include "stdb/query/steps/aggregate_processing_step.h"
+#include "stdb/query/steps/chain.h"
+#include "stdb/query/steps/filter_processing_step.h"
+#include "stdb/query/steps/group_aggregate_combiner.h"
+#include "stdb/query/steps/group_aggregate_filter_processing_step.h"
+#include "stdb/query/steps/group_aggregate_processing_step.h"
+#include "stdb/query/steps/join.h"
+#include "stdb/query/steps/materialization_step.h"
+#include "stdb/query/steps/mergeby.h"
+#include "stdb/query/steps/processing_prelude.h"
+#include "stdb/query/steps/scan_events_processing_step.h"
+#include "stdb/query/steps/scan_processing_step.h"
+#include "stdb/query/steps/series_order_aggregate.h"
+#include "stdb/query/steps/time_order_aggregate.h"
 
 namespace stdb {
 namespace qp {
