+++ stdb/query/steps/aggregate.h
@@ -1,10 +1,10 @@
 /*!
  * \file aggregate.h
  */
-#ifndef STDB_QUERY_PLAN_AGGREGATE_H_
-#define STDB_QUERY_PLAN_AGGREGATE_H_
+#ifndef STDB_QUERY_STEPS_AGGREGATE_H_
+#define STDB_QUERY_STEPS_AGGREGATE_H_
 
-#include "stdb/query/plan/materialization_step.h"
+#include "stdb/query/steps/materialization_step.h"
 
 namespace stdb {
 namespace qp {
@@ -53,4 +53,4 @@ struct Aggregate : MaterializationStep {
 }  // namespace qp
 }  // namespace stdb
 
-#endif  // STDB_QUERY_PLAN_AGGREGATE_H_
+#endif  // STDB_QUERY_STEPS_AGGREGATE_H_
