+++ stdb/query/steps/group_aggregate_combiner.h
@@ -1,10 +1,10 @@
 /*!
  * \file group_aggregate_combiner.h
  */
-#ifndef STDB_QUERY_PLAN_GROUP_AGGREGATER_COMBINER_H_
-#define STDB_QUERY_PLAN_GROUP_AGGREGATER_COMBINER_H_
+#ifndef STDB_QUERY_STEPS_GROUP_AGGREGATER_COMBINER_H_
+#define STDB_QUERY_STEPS_GROUP_AGGREGATER_COMBINER_H_
 
-#include "stdb/query/plan/materialization_step.h"
+#include "stdb/query/steps/materialization_step.h"
 
 namespace stdb {
 namespace qp {
@@ -99,4 +99,4 @@ struct GroupAggregateCombiner : MaterializationStep {
 }  // namespace qp
 }  // namespace stdb
 
-#endif  // STDB_QUERY_PLAN_GROUP_AGGREGATER_COMBINER_H_
+#endif  // STDB_QUERY_STEPS_GROUP_AGGREGATER_COMBINER_H_
