+++ stdb/query/steps/group_aggregate_processing_step.h
@@ -1,10 +1,10 @@
 /*!
  * \file group_aggregate_processing_step.h
  */
-#ifndef STDB_QUERY_PLAN_GROUP_AGGREGATE_PROCESSING_STEP_H_
-#define STDB_QUERY_PLAN_GROUP_AGGREGATE_PROCESSING_STEP_H_
+#ifndef STDB_QUERY_STEPS_GROUP_AGGREGATE_PROCESSING_STEP_H_
+#define STDB_QUERY_STEPS_GROUP_AGGREGATE_PROCESSING_STEP_H_
 
-#include "stdb/query/plan/processing_prelude.h"
+#include "stdb/query/steps/processing_prelude.h"
 
 namespace stdb {
 namespace qp {
@@ -102,4 +102,4 @@ struct GroupAggregateProcessingStep : ProcessingPrelude {
 }  // namespace qp
 }  // namespace stdb
 
-#endif  // STDB_QUERY_PLAN_GROUP_AGGREGATE_PROCESSING_STEP_H_
+#endif  // STDB_QUERY_STEPS_GROUP_AGGREGATE_PROCESSING_STEP_H_
