+++ stdb/query/steps/mergeby.h
@@ -1,10 +1,10 @@
 /*!
  * \file mergeby.h
  */
-#ifndef STDB_QUERY_PLAN_MERGEBY_H_
-#define STDB_QUERY_PLAN_MERGEBY_H_
+#ifndef STDB_QUERY_STEPS_MERGEBY_H_
+#define STDB_QUERY_STEPS_MERGEBY_H_
 
-#include "stdb/query/plan/materialization_step.h"
+#include "stdb/query/steps/materialization_step.h"
 
 namespace stdb {
 namespace qp {
@@ -92,4 +92,4 @@ struct MergeBy : MaterializationStep {
 }  // namespace qp
 }  // namespace stdb
 
-#endif  // STDB_QUERY_PLAN_MERGEBY_H_
+#endif  // STDB_QUERY_STEPS_MERGEBY_H_
