+++ stdb/query/steps/scan_events_processing_step.h
@@ -1,10 +1,10 @@
 /*!
  * \file scan_events_processing_step.h
  */
-#ifndef STDB_QUERY_PLAN_SCAN_EVENTS_PROCESSING_STEP_H_
-#define STDB_QUERY_PLAN_SCAN_EVENTS_PROCESSING_STEP_H_
+#ifndef STDB_QUERY_STEPS_SCAN_EVENTS_PROCESSING_STEP_H_
+#define STDB_QUERY_STEPS_SCAN_EVENTS_PROCESSING_STEP_H_
 
-#include "stdb/query/plan/processing_prelude.h"
+#include "stdb/query/steps/processing_prelude.h"
 
 namespace stdb {
 namespace qp {
@@ -64,4 +64,4 @@ struct ScanEventsProcessingStep : ProcessingPrelude {
 }  // namespace qp
 }  // namespace stdb
 
-#endif  // STDB_QUERY_PLAN_SCAN_EVENTS_PROCESSING_STEP_H_
+#endif  // STDB_QUERY_STEPS_SCAN_EVENTS_PROCESSING_STEP_H_
